Move generator and descriptor back to public API
These two packages are depended on by at least one
other popular package.
