+++ protoc-gen-grpc-gateway/descriptor/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
         "services.go",
         "types.go",
     ],
-    importpath = "github.com/grpc-ecosystem/grpc-gateway/internal/descriptor",
+    importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway/descriptor",
     deps = [
         "//internal/httprule:go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
