+++ protoc-gen-grpc-gateway/generator/BUILD.bazel
@@ -5,9 +5,9 @@ package(default_visibility = ["//visibility:public"])
 go_library(
     name = "go_default_library",
     srcs = ["generator.go"],
-    importpath = "github.com/grpc-ecosystem/grpc-gateway/internal/generator",
+    importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway/generator",
     deps = [
-        "//internal/descriptor:go_default_library",
+        "//protoc-gen-grpc-gateway/descriptor:go_default_library",
         "@io_bazel_rules_go//proto/wkt:compiler_plugin_go_proto",
     ],
 )
