+++ protoc-gen-grpc-gateway/internal/gengateway/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     ],
     importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway/internal/gengateway",
     deps = [
-        "//internal/descriptor:go_default_library",
-        "//internal/generator:go_default_library",
+        "//protoc-gen-grpc-gateway/descriptor:go_default_library",
+        "//protoc-gen-grpc-gateway/generator:go_default_library",
         "//utilities:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
@@ -30,8 +30,8 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//internal/descriptor:go_default_library",
         "//internal/httprule:go_default_library",
+        "//protoc-gen-grpc-gateway/descriptor:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
     ],
