+++ protoc-gen-grpc-gateway/internal/gengateway/generator_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/golang/protobuf/proto"
 	protodescriptor "github.com/golang/protobuf/protoc-gen-go/descriptor"
-	"github.com/grpc-ecosystem/grpc-gateway/internal/descriptor"
+	"github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway/descriptor"
 )
 
 func newExampleFileDescriptor() *descriptor.File {
