+++ protoc-gen-swagger/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger",
     deps = [
         "//internal/codegenerator:go_default_library",
-        "//internal/descriptor:go_default_library",
+        "//protoc-gen-grpc-gateway/descriptor:go_default_library",
         "//protoc-gen-swagger/internal/genswagger:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
