+++ protoc-gen-swagger/internal/genswagger/BUILD.bazel
@@ -15,8 +15,8 @@ go_library(
     importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger/internal/genswagger",
     deps = [
         "//internal:go_default_library",
-        "//internal/descriptor:go_default_library",
-        "//internal/generator:go_default_library",
+        "//protoc-gen-grpc-gateway/descriptor:go_default_library",
+        "//protoc-gen-grpc-gateway/generator:go_default_library",
         "//protoc-gen-swagger/options:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//descriptor:go_default_library_gen",
@@ -36,8 +36,8 @@ go_test(
     srcs = ["template_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//internal/descriptor:go_default_library",
         "//internal/httprule:go_default_library",
+        "//protoc-gen-grpc-gateway/descriptor:go_default_library",
         "//protoc-gen-swagger/options:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@io_bazel_rules_go//proto/wkt:any_go_proto",
